modal - create a modal container for flexible styling
