+++ src/index.js
@@ -38,6 +38,7 @@ export { default as WidgetStats } from './widget-stats'
 export { default as withRef } from './with-ref'
 export { TabPanelsWithRef }
 export { default as Modal } from './modal'
+export { default as ModalContainer } from './modal-container'
 
 // deprecated
 export { default as DataTable } from './deprecated/data-table' // deprecated by Table
